Feature/picklejar (#3)
* pickleJar.py : added basic save / load functions. Added functions to convert sqlite dbs to dataDicts and pickle them

* pickleJar.py
tested and debugged applyFunctionToData, applyFunctionToDir

* pickleJar.py
added scan plotting, pixel plotting, and associated data manipulation functions. most of sqliteUtils functionality now works with pickles

* pickleJar.py
added baselineCorrectionScans for scans with additive baseline drift and multiplicative changes in span
added waveform plotting functions

* runUltrasonicExperiment: added pickleData option, which will automatically convert sqlite data into a pickled python dict
Added a call to sqliteToPickle at the end of ultrasonicScan and repeatPulse if the data is saved as sqlite and pickleData is True

* Added first version of exampleAnalysis.py, which walks through the use of the analysis and plotting functions in pickleJar.py
These examples still need to be tested. It will probably be useful to include some example data in the git

* Edited and debugged the examples in exampleAnalysis.py

Added a function pj.writeDataToDict() to write a constant to a dataDict
Updated generateScanPlots to restore the previous matplotlib backend after it finishes saving plots
Added error handling to sqliteToPickle to check if the pickle file already exists
Renamed listPicklesInDirectory to listFilesInDirectory. Now takes an optional second argument to specify the file extension, defaulting to .pickle
Added staltaFirstBreak function to list of analysis functions
Fixed a bug in multiScanDataAtPixels that was failing to collect non-float data
Added xDat argument to waveform plotting functions so that the x-axis data can be changed from time (for plotting e.g. fft)

Changed some comments in database.py and sqliteUtils.py

* Added absoluteSum() to pickleJar analysis functions
Added current analysis functions and usage in the comments at the end of exampleAnalysis.py

* Fixed implementation of sqlite pickling in repeatPulse experiment