Large overhaul to picosdkRapidblockPulse.py: renamed class to Picoscope (refactored in experiment functions too)
Added collectionDirection parameter which determines which channel is the pulser and which is the transmission

Changed all Picoscope functions to save experiment parameters as class variable to simplify inputs

Renamed runPicoMeasurement to runRapidBlock. Made a new runPicoMeasurement which acts as a wrapper to runRapidBlock and sorts the outputs into a dict to enable arbitrary modes and directions with clean outputs

Moved setupPicoMeasurement to be a helper function for runRapidBlock(). Changed runRapidBlock and setupPicoMeasurement to only take 'direction' as an input (this enables runPicoMeasurement to execute 'forward' and 'reverse' measurements when the user inputs 'both' as the direction)

Updated voltageRangeFinder and singlePulseMeasure to work with the overhauled code

Created helper functions voltageIndexFromRange, setChannels, transmissionMaximum, and plotWaveDict to simplify code